requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astronomer-cosmos==1.0.4
apache-airflow-providers-microsoft-mssql==3.8.0
from airflow.providers.microsoft.mssql.operators.mssql import MsSqlOperator